Big mess of module renaming, declarative_charts.js was a very outdated name and we didn't need a file exporting both A11yTable and GraphLegend
