+++ client/src/charts/shared.js
@@ -2,13 +2,13 @@ import { businessConstants } from '../models/businessConstants.js';
 import { NA_color } from '../core/color_schemes.js';
 import { formats } from "../core/format.js";
 
+
 export const infobase_colors_smart = (col_scale) => (label) => {
   if ( _.includes(businessConstants.NA_values,label) ){
     return NA_color;
   }
   return col_scale(label);
 };
-
 
 export const get_formatter = (is_money, formatter, raw = true, full = false) => (
   _.isUndefined(formatter) ?
