+++ client/src/panels/panel_declarations/finances/auth_exp_planned_spending/auth_exp_planned_spending.js
@@ -7,7 +7,8 @@ import {
   run_template,
   year_templates,
   actual_to_planned_gap_year,
-  declarative_charts,
+  GraphLegend,
+  A11yTable,
   StdPanel,
   Col,
   create_text_maker_component,
@@ -18,10 +19,6 @@ import {
   declare_panel,
 } from "../../shared.js";
 
-const { 
-  GraphLegend,
-  A11YTable,
-} = declarative_charts;
 const { Details } = util_components;
 
 const { std_years, planning_years, estimates_years } = year_templates;
@@ -57,7 +54,7 @@ const AuthExpPlannedSpendingTable = ({data_series}) => {
   );
   
   return (
-    <A11YTable
+    <A11yTable
       data_col_headers={series_labels}
       data={data}
     />
