+++ client/src/panels/panel_declarations/people/employee_fol.js
@@ -5,7 +5,7 @@ import {
   businessConstants,
   year_templates,
   create_text_maker_component,
-  declarative_charts,
+  A11yTable,
   StdPanel,
   Col,
 
@@ -19,8 +19,6 @@ const { text_maker, TM } = create_text_maker_component(text);
 
 const { people_years } = year_templates;
 const { fol } = businessConstants;
-
-const { A11YTable } = declarative_charts;
 
 
 const info_deps_by_level = {
@@ -121,7 +119,7 @@ export const declare_employee_fol_panel = () => declare_panel({
           }
           { window.is_a11y_mode &&
             <Col size={12} isGraph>
-              <A11YTable
+              <A11yTable
                 label_col_header = {text_maker("FOL")}
                 data_col_headers = {[...ticks, text_maker("five_year_percent_header")]}
                 data = {_.map(panel_args, dimension => {
