+++ client/src/panels/panel_declarations/people/employee_gender.js
@@ -5,7 +5,7 @@ import {
   businessConstants,
   year_templates,
   create_text_maker_component,
-  declarative_charts,
+  A11yTable,
   StdPanel,
   Col,
 
@@ -18,8 +18,6 @@ const { text_maker, TM } = create_text_maker_component(text);
 
 const { people_years } = year_templates;
 const { gender } = businessConstants;
-
-const { A11YTable } = declarative_charts;
 
 
 const info_deps_by_level = {
@@ -120,7 +118,7 @@ export const declare_employee_gender_panel = () => declare_panel({
           }
           { window.is_a11y_mode &&
             <Col size={12} isGraph>
-              <A11YTable
+              <A11yTable
                 label_col_header = {text_maker("employee_gender")}
                 data_col_headers = {[...ticks, text_maker("five_year_percent_header")]}
                 data = {_.map(panel_args, dimension => {
